Refactor new task commands into same superclass
To reduce repetition of code, Event, Deadline and ToDo creation command
classes now inherit from a common NewTaskCommand.
